% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANYSXP.R
\name{ANYSXP}
\alias{ANYSXP}
\title{The type of the "any" object}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{ANYSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of "any" objects.
}
\section{Declaration}{
\preformatted{#define ANYSXP 18 /* make "any" args work.
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
