% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCODESXP.R
\name{BCODESXP}
\alias{BCODESXP}
\title{The type of byte code objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{BCODESXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of byte code objects.
}
\section{Declaration}{
\preformatted{#define BCODESXP 21 /* byte code */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
