% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILTINSXP.R
\name{BUILTINSXP}
\alias{BUILTINSXP}
\title{The type of builtin non-special forms}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{BUILTINSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of builtin non-special forms.
}
\section{Declaration}{
\preformatted{#define BUILTINSXP 8 /* builtin non-special forms */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
