\name{C3PO-package}
\alias{C3PO-package}
\alias{C3PO}
\docType{package}
\title{
\packageTitle{C3PO}
}
\description{
The \code{C3PO} package provides the function \code{\link{r2d2}}, implemented in C,  to perform
R2D2 multivariate bias correction introduced in Vrac et al. (2019)
}
\author{
\packageAuthor{C3PO}

Maintainer: \packageMaintainer{C3PO}
}

\references{
\emph{Vrac, M.
Multivariate bias adjustment of high-dimensional climate simulations:
the Rank Resampling for Distributions and Dependences
(R2D2) Bias Correction. Hydrol. Earth Syst. Sci., 22, 3175-3196},
\url{https://doi.org/10.5194/hess-22-3175-2018}
}
\keyword{ package }
\seealso{
\code{\link[R2D2:R2D2-package]{R2D2}} ~~
}

