% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHARACTER_VALUE.R
\name{CHARACTER_VALUE}
\alias{CHARACTER_VALUE}
\alias{STRING_VALUE}
\title{Convert a scalar R object to a C string}
\arguments{
\item{x}{a pointer \code{\link{SEXP}}, of which only the first element is
used.}
}
\value{
C string, i.e., a pointer \code{char *}.
}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mfn')} \Sexpr[results=rd, stage=render]{c3po:::badge('r2c')}

Both \code{CHARACTER_VALUE} and \code{STRING_VALUE} convert a scalar R object to a
C-string \code{(char *)}.
}
\section{Declaration}{
\preformatted{#define CHARACTER_VALUE(x) CHAR(Rf_asChar(x))
#define STRING_VALUE(x) CHAR(Rf_asChar(x))
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\examples{
# Convert an R object to a C-string (char *)
to_string <- inline::cfunction(c(x = "any"),
' const char* x_;
  x_ = CHARACTER_VALUE(x);
  Rprintf("\%s\\\\n", x_);
  return R_NilValue;
')

invisible(to_string(1L))
invisible(to_string(pi))
invisible(to_string(TRUE))
invisible(to_string(list(a = 'b')))

}
\seealso{
\link{R_CHAR},
\link{Rf_asChar}
}
