% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHARSXP.R
\name{CHARSXP}
\alias{CHARSXP}
\title{The type of scalar strings (internal only)}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{CHARSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of "scalar" string objects.
}
\section{Declaration}{
\preformatted{#define CHARSXP 9 /* "scalar" string type (internal only)*/
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
