% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLOSXP.R
\name{CLOSXP}
\alias{CLOSXP}
\title{The type of function objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{CLOSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of functions, i.e., \code{\link{closure}} objects.
}
\section{Declaration}{
\preformatted{#define CLOSXP 3 /* closures */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
