% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPLXSXP.R
\name{CPLXSXP}
\alias{CPLXSXP}
\title{The type of complex vectors}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{CPLXSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{complex}} vectors.
}
\section{Declaration}{
\preformatted{#define CPLXSXP 15 /* complex variables */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
