% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOTSXP.R
\name{DOTSXP}
\alias{DOTSXP}
\title{The type of \code{...} object}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{DOTSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{dots}} (\code{...}) object.
}
\section{Declaration}{
\preformatted{#define DOTSXP 17 /* dot-dot-dot object */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
