% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENVSXP.R
\name{ENVSXP}
\alias{ENVSXP}
\title{The type of environment objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{ENVSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{environment}} objects.
}
\section{Declaration}{
\preformatted{#define ENVSXP 4 /* environments */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
