% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPRSXP.R
\name{EXPRSXP}
\alias{EXPRSXP}
\title{The type of expression objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{EXPRSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{expression}} objects.
}
\section{Declaration}{
\preformatted{#define EXPRSXP 20 /* expressions vectors */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
