% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXTPTRSXP.R
\name{EXTPTRSXP}
\alias{EXTPTRSXP}
\title{The type of external pointers}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{EXTPTRSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of external pointer objects.
}
\section{Declaration}{
\preformatted{#define EXTPTRSXP 22 /* external pointer */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
