% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FREESXP.R
\name{FREESXP}
\alias{FREESXP}
\title{The type of R's node released by GC}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{FREESXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of nodes released by garbage collector.
}
\section{Declaration}{
\preformatted{#define FREESXP 31 /* node released by GC */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
