% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNSXP.R
\name{FUNSXP}
\alias{FUNSXP}
\title{The pseudo-type lumping closure, builtin and special}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{FUNSXP} is an object-like macro defining a \code{\link{FUNSXP}}
pseudo-type of lumping closure, builtin and special objects.
}
\section{Declaration}{
\preformatted{#define FUNSXP 99 /* Closure or Builtin or Special */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
