% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTSXP.R
\name{INTSXP}
\alias{INTSXP}
\title{The type of integer vectors}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{INTSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{integer}} vectors.
}
\section{Declaration}{
\preformatted{#define INTSXP 13 /* integer vectors */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
