% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANGSXP.R
\name{LANGSXP}
\alias{LANGSXP}
\title{The type of language constructs}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{LANGSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of language constructs such \code{\link{call}} and \code{\link{formula}}
objects.
}
\section{Declaration}{
\preformatted{#define LANGSXP 6 /* language constructs (special lists) */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
