% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGLSXP.R
\name{LGLSXP}
\alias{LGLSXP}
\title{The type of logical vectors}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{LGLSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{logical}} vectors.
}
\section{Declaration}{
\preformatted{#define LGLSXP 10 /* logical vectors */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
