% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LISTSXP.R
\name{LISTSXP}
\alias{LISTSXP}
\title{The type of pairlist objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{LISTSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{pairlist}} objects.
}
\section{Declaration}{
\preformatted{#define LISTSXP 2 /* lists of dotted pairs */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
