% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEWSXP.R
\name{NEWSXP}
\alias{NEWSXP}
\title{The type of fresh node created in new page}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{NEWSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of fresh node objects created in new page.
}
\section{Declaration}{
\preformatted{#define NEWSXP 30 /* fresh node created in new page */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
