% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NILSXP.R
\name{NILSXP}
\alias{NILSXP}
\title{NILSXP}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

There is only one object of type NILSXP, \link{R_NilValue}, with no data.
}
\section{Declaration}{
\preformatted{define NILSXP 0 /* nil = NULL */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{R_NilValue}
}
