% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROMSXP.R
\name{PROMSXP}
\alias{PROMSXP}
\title{The type of promise objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{PROMSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{promise}} objects.
}
\section{Declaration}{
\preformatted{#define PROMSXP 5 /* promises: [un]evaluated closure arguments */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
