% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_CHAR.R
\name{R_CHAR}
\alias{R_CHAR}
\alias{CHAR}
\title{Convert a string scalar (CHARSXP) to a C-string (char *)}
\arguments{
\item{x}{a pointer \code{\link{SEXP}}, of type \code{\link{CHARSXP}}.}
}
\value{
C string, i.e., a pointer \code{char *}.
}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('fnp')} \Sexpr[results=rd, stage=render]{c3po:::badge('r2c')}

Convert a string scalar (\code{\link{CHARSXP}}) to a C-string (\code{char
*}).
}
\section{Declaration}{
\preformatted{const char *(R_CHAR)(SEXP x);
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\section{Definition}{
\preformatted{const char *(R_CHAR)(SEXP x) \{
  if(TYPEOF(x) != CHARSXP) // Han-Tak proposes to prepend  'x && '
  error("\%s() can only be applied to a '\%s', not a '\%s'",
        "CHAR", "CHARSXP", type2char(TYPEOF(x)));
  return (const char *) CHAR(CHK(x));
\}
}

In \href{https://github.com/wch/r-source/blob/trunk/src/main/memory.c}{memory.c}.
}

\seealso{
\link{Rf_asChar}
}
