% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_NilValue.R
\name{R_NilValue}
\alias{R_NilValue}
\title{R_NilValue}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('variable')}

\code{R_NilValue} is a global (external) variable of type \code{\link{SEXP}}.
\code{R_NilValue} points to the C level implementation of R's
\code{\link{NULL}} object.
}
\section{Declaration}{
\preformatted{LibExtern SEXP	R_NilValue; /* The nil object */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\section{Definition}{
\preformatted{GET_FREE_NODE(R_NilValue);
}

In \href{https://github.com/wch/r-source/blob/trunk/src/main/memory.c}{memory.c}.
}

\seealso{
\link{NULL},
\link{NILSXP},
\link{TYPEOF}
}
