% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAWSXP.R
\name{RAWSXP}
\alias{RAWSXP}
\title{The type of raw vectors}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{RAWSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{raw}} vectors.
}
\section{Declaration}{
\preformatted{#define RAWSXP 24 /* raw bytes */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
