% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbyte.R
\name{Rbyte}
\alias{Rbyte}
\title{Rbyte}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('typedef-dtype')}

Rbyte is an alias to an \code{unsigned char}.
}
\section{Declaration}{
\preformatted{typedef unsigned char Rbyte;
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

