% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REALSXP.R
\name{REALSXP}
\alias{REALSXP}
\title{The type of double vectors}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{REALSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{double}} vectors.
}
\section{Declaration}{
\preformatted{#define REALSXP 14 /* real variables */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
