% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rf_ScalarInteger.R
\name{Rf_ScalarInteger}
\alias{Rf_ScalarInteger}
\alias{ScalarInteger}
\title{Rf_ScalarInteger}
\arguments{
\item{x}{an int value.}
}
\value{
A \link{SEXP} object, namely a length one \code{integer} vector.
}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('fnp')} \Sexpr[results=rd, stage=render]{c3po:::badge('c2r')} \Sexpr[results=rd, stage=render]{c3po:::badge('prot')}

Converts an \code{int x} to a \link{SEXP} object that represents an \R length one
\code{integer} vector.
}
\section{Declaration}{
\preformatted{SEXP Rf_ScalarInteger(int x);
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\section{Definition}{
\preformatted{INLINE_FUN SEXP ScalarInteger(int x)
\{
  SEXP ans = allocVector(INTSXP, 1);
  SET_SCALAR_IVAL(ans, x);
  return ans;
\}
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinlinedfuns.h}{Rinlinedfuns.h}.
}

\examples{
# Convert a C int number as an R length one integer vector
forty_two <- inline::cfunction(NULL,
 ' int x = 42;
   SEXP y;
   y = PROTECT(Rf_ScalarInteger(x));
   UNPROTECT(1);
   return y;
 ')
forty_two()


}
