% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rf_ScalarRaw.R
\name{Rf_ScalarRaw}
\alias{Rf_ScalarRaw}
\alias{ScalarRaw}
\title{Rf_ScalarRaw}
\arguments{
\item{x}{an \link{Rbyte} value.}
}
\value{
A \link{SEXP} object, namely a length one (one byte) \code{raw} vector.
}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('fnp')} \Sexpr[results=rd, stage=render]{c3po:::badge('c2r')} \Sexpr[results=rd, stage=render]{c3po:::badge('prot')}

Converts an \link{Rbyte} (unsigned char) to a \link{SEXP} object that represents an R
length one \code{raw} vector.
}
\section{Declaration}{
\preformatted{SEXP Rf_ScalarRaw(Rbyte x);
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\section{Definition}{
\preformatted{INLINE_FUN SEXP ScalarRaw(Rbyte x)
\{
  SEXP ans = allocVector(RAWSXP, 1);
  SET_SCALAR_BVAL(ans, x);
  return ans;
\}
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinlinedfuns.h}{Rinlinedfuns.h}.
}

\examples{
# Convert a C unsigned char to an R length one raw vector
asterisk_byte <- inline::cfunction(NULL,
 " Rbyte x = '*'; // '*' is 42 in decimal, 2a in hexadecimal.
   SEXP y;
   y = PROTECT(Rf_ScalarRaw(x));
   UNPROTECT(1);
   return y;
 ")
asterisk_byte()

}
