% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rinternals.R
\name{Rinternals.h}
\alias{Rinternals.h}
\title{Rinternals.h}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('cheader')}

Rinternals.h header file.
}
\section{Path}{

/usr/include/R/Rinternals.h
}

