% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXPREC.R
\name{SEXPREC}
\alias{SEXPREC}
\title{R's fundamental data type at C level}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('typedef-str')}\Sexpr[results=rd, stage=render]{c3po:::badge('USE_RINTERNALS')}

S-expression (SEXP) record (REC) structure, a C structure underlying every R
object. Accessible via a pointer of type \code{\link{SEXP}}.
\code{\link{SEXPREC}} is a variant type, with subtypes for all \R’s data
structures (see \code{\link{SEXPTYPE}} for a list of subtypes).

The SEXPREC is composed of two parts (five struct members in total):
\itemize{
\item Header, \code{\link{SEXPREC_HEADER}}:
\itemize{
\item \code{struct \link{sxpinfo_struct} sxpinfo}
\item \code{struct SEXPREC *attrib}
\item \code{struct SEXPREC *gengc_next_node}
\item \code{struct SEXPREC *gengc_prev_node}
}
}
\itemize{
\item Node data, \code{u}.
}
}
\section{Declaration}{
\preformatted{typedef struct SEXPREC \{
    SEXPREC_HEADER;
    union \{
  struct primsxp_struct primsxp;
  struct symsxp_struct symsxp;
  struct listsxp_struct listsxp;
  struct envsxp_struct envsxp;
  struct closxp_struct closxp;
  struct promsxp_struct promsxp;
    \} u;
\} SEXPREC;
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#The-_0027data_0027}
}
\seealso{
\link{SEXP}
}
