% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXPREC_HEADER.R
\name{SEXPREC_HEADER}
\alias{SEXPREC_HEADER}
\title{SEXPREC_HEADER}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')} \Sexpr[results=rd, stage=render]{c3po:::badge('USE_RINTERNALS')}

SEXPREC_HEADER is a macro comprising the declaration of four objects:
\itemize{
\item \code{struct \link{sxpinfo_struct} sxpinfo}
\item \code{struct \link{SEXPREC} *attrib}
\item \code{struct \link{SEXPREC} *gengc_next_node}
\item \code{struct \link{SEXPREC} *gengc_prev_node}
}

SEXPREC_HEADER is used in the declaration of \link{SEXPREC}.
}
\section{Declaration}{
\preformatted{#define SEXPREC_HEADER \\
    struct sxpinfo_struct sxpinfo; \\
    struct SEXPREC *attrib; \\
    struct SEXPREC *gengc_next_node, *gengc_prev_node;
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\seealso{
\link{SEXPREC},
\link{sxpinfo_struct}
}
