% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIALSXP.R
\name{SPECIALSXP}
\alias{SPECIALSXP}
\title{The type of special forms}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{SPECIALSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of special form objects.
}
\section{Declaration}{
\preformatted{#define SPECIALSXP 7 /* special forms */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
