% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRSXP.R
\name{STRSXP}
\alias{STRSXP}
\title{The type of character vectors}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{STRSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{character}} vectors.
}
\section{Declaration}{
\preformatted{#define STRSXP 16 /* string vectors */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
