% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpinfo_struct.R
\name{sxpinfo_struct}
\alias{sxpinfo_struct}
\title{struct sxpinfo_struct}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('struct')} \Sexpr[results=rd, stage=render]{c3po:::badge('USE_RINTERNALS')}

\code{struct sxpinfo_struct} is a struct of 13 members:
\itemize{
\item \link{SEXPTYPE}\code{ type}
\item \verb{unsigned int scalar}
\item \verb{unsigned int obj}
\item \verb{unsigned int alt}
\item \verb{unsigned int gp}
\item \verb{unsigned int mark}
\item \verb{unsigned int debug}
\item \verb{unsigned int trace}
\item \verb{unsigned int spare}
\item \verb{unsigned int gcgen}
\item \verb{unsigned int gccls}
\item \verb{unsigned int named}
\item \verb{unsigned int extra}
}

\code{struct sxpinfo_struct} is used in the definition of \link{SEXPREC_HEADER}.
}
\section{Declaration}{
\preformatted{struct sxpinfo_struct \{
  SEXPTYPE type      :  TYPE_BITS;
  /* ==> (FUNSXP == 99) \%\% 2^5 == 3 == CLOSXP
   * -> warning: `type' is narrower than values
   *              of its type
   * when SEXPTYPE was an enum */
  unsigned int scalar:  1;
  unsigned int obj   :  1;
  unsigned int alt   :  1;
  unsigned int gp    : 16;
  unsigned int mark  :  1;
  unsigned int debug :  1;
  unsigned int trace :  1;  /* functions and memory tracing */
  unsigned int spare :  1;  /* used on closures and when REFCNT is defined */
  unsigned int gcgen :  1;  /* old generation number */
  unsigned int gccls :  3;  /* node class */
  unsigned int named : NAMED_BITS;
  unsigned int extra : 32 - NAMED_BITS; /* used for immediate bindings */
\}; /*     Tot: 64 */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\seealso{
\link{SEXPTYPE},
\link{SEXPREC_HEADER},
\link{SEXPREC},
}
