% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYMSXP.R
\name{SYMSXP}
\alias{SYMSXP}
\title{The type of name and symbol objects}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{SYMSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of \code{\link{name}} and symbol objects.
}
\section{Declaration}{
\preformatted{#define SYMSXP 1 /* symbols */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
