% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VECSXP.R
\name{VECSXP}
\alias{VECSXP}
\title{The type of generic vectors (lists)}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{VECSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of generic vectors, i.e., \code{\link{list}} objects.
}
\section{Declaration}{
\preformatted{#define VECSXP	    19	  /* generic vectors */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
