% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEAKREFSXP.R
\name{WEAKREFSXP}
\alias{WEAKREFSXP}
\title{The type of weak references}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('mobj')}

\code{WEAKREFSXP} is an object-like macro defining the \code{\link{SEXPTYPE}}
type of weak reference objects.
}
\section{Declaration}{
\preformatted{#define WEAKREFSXP 23 /* weak reference */
}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{SEXPTYPE},
\link{TYPEOF}
}
