% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3po-package.R
\docType{package}
\name{c3po-package}
\alias{c3po}
\alias{c3po-package}
\title{c3po: From C to R and back again}
\description{
Non-official documentation for R's C internal API.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Centre for Biomedical Research, University of Algarve [copyright holder, funder]
}

}
\keyword{internal}
