% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEXPTYPE.R
\name{SEXPTYPE}
\alias{SEXPTYPE}
\title{SEXPTYPE}
\description{
\Sexpr[results=rd, stage=render]{c3po:::badge('typedef-str')} \Sexpr[results=rd, stage=render]{c3po:::badge('USE_RINTERNALS')}

SEXPTYPE is an alias to an \code{unsigned int}. It is used to encode the type
of R object that a SEXPREC holds, namely in \code{struct
\link{sxpinfo_struct}}, where it works to declare the type of the struct
member \code{type}.
}
\section{Declaration}{
\preformatted{}

In \href{https://github.com/wch/r-source/blob/trunk/src/include/Rinternals.h}{Rinternals.h}.
}

\references{
\url{https://cran.r-project.org/doc/manuals/r-patched/R-ints.html#SEXPTYPEs}
}
\seealso{
\link{sxpinfo_struct},
\link{SEXP},
\link{SEXPREC}
}
